+++ src/components/Autocomplete/Autocomplete.types.ts
@@ -1,9 +1,9 @@
 import type { TextFieldProps } from "components/TextField";
 import type { MenuItemProps } from "components/MenuItem";
-import type { DropdownMenuProps } from "components/DropdownMenu";
+import type { DropdownMenuProps, DropdownMenuInstance } from "components/DropdownMenu";
 
 export type Props = TextFieldProps &
-	Pick<DropdownMenuProps, "containerRef"> & {
+	Pick<DropdownMenuProps, "containerRef" | "instanceRef"> & {
 		onInput?: TextFieldProps["onChange"];
 		onItemSelect?: (args: { value: string }) => void;
 		onBackspace?: () => void;
@@ -15,3 +15,5 @@ export type ItemProps = MenuItemProps & { value: string };
 export type Context = {
 	onItemClick: (args: { value: string }) => void;
 };
+
+export type Instance = DropdownMenuInstance;