+++ src/components/Autocomplete/index.ts
@@ -1,2 +1,5 @@
 export { default } from "./Autocomplete";
-export type { Props as AutocompleteProps } from "./Autocomplete.types";
+export type {
+	Props as AutocompleteProps,
+	Instance as AutocompleteInstance,
+} from "./Autocomplete.types";