+++ src/components/Autocomplete/tests/Autocomplete.stories.tsx
@@ -132,6 +132,7 @@ export const multiselect = () => {
 	return (
 		<>
 			<Autocomplete
+				instanceRef={ref}
 				name="fruit"
 				value={query}
 				placeholder="Pick your food"